Migrate to src/ layout
The module now includes the very basic artefacts that could be used
to install with pip, as step 1 towards getting it onto pypi.

I also removed some things that really shouldn't be there, such as
the __pycache__ directory, and updated .gitignore to make sure they
don't come back.

This is still by no means best practice, but I think it's taking a
step in more or less the right direction.